Remove CI build job — Vercel handles production builds
Clerk rejects placeholder keys during Next.js prerendering, causing
CI build failures. Since lint, type-check, and tests already catch
code issues, the build step with fake env vars is unnecessary.
