+++ .github/workflows/ci.yml
@@ -54,39 +54,3 @@ jobs:
 
       - name: Unit tests
         run: npm run test
-
-  build:
-    name: Build
-    runs-on: ubuntu-latest
-    timeout-minutes: 15
-    needs: [lint, test]
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 1
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 22
-          cache: npm
-
-      - run: npm ci
-
-      - name: Build Next.js
-        run: npm run build
-        env:
-          # Client env vars (NEXT_PUBLIC_*)
-          NEXT_PUBLIC_CONVEX_URL: https://ci-placeholder.convex.cloud
-          NEXT_PUBLIC_CLERK_FRONTEND_API_URL: https://ci-placeholder.clerk.accounts.dev
-          NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY: pk_test_ci_placeholder
-          NEXT_PUBLIC_CINETPAY_SITE_ID: ci_placeholder
-          NEXT_PUBLIC_CINETPAY_API_KEY: ci_placeholder
-          # Server env vars (required by Zod validation at build time)
-          CONVEX_DEPLOYMENT: ci-placeholder
-          CLERK_SECRET_KEY: sk_test_ci_placeholder
-          CLERK_WEBHOOK_SECRET: whsec_ci_placeholder
-          STRIPE_SECRET_KEY: sk_test_ci_placeholder
-          STRIPE_WEBHOOK_SECRET: whsec_ci_placeholder
-          STRIPE_PRICE_ID: price_ci_placeholder
-          RESEND_API_KEY: re_ci_placeholder
-          CINETPAY_SECRET_KEY: ci_placeholder
